Rice Disease Detection/requirements.txt
Flask==2.3.3
tensorflow>=2.13.0,<2.17.0
numpy>=1.21.0,<2.0.0
Pillow>=9.0.0,<11.0.0
scikit-learn>=1.3.0,<2.0.0
matplotlib>=3.7.0,<4.0.0
keras>=2.13.0,<2.17.0
Werkzeug==2.3.7
h5py>=3.8.0 